3rd table working the DELETE and GET, CREATE Was able to nest the handlebars togethere to get the GET. Able to get DELETE working.  Need to do update and then backend will be completed. Need to work on stage 2 handlebars to make forms look nice then will produce to login change password.
NOTES: Still using prompts for update / edit on LIST.  Need to change it
to dialog box but need to review get the ID outside of the handle to
work on a that dialog prompt box that i created.

Will work on this and after should be able to the task update list.